chore(types): move remaining *StyleProps types to components
