+++ packages/fluentui/react/src/themes/teams/types.ts
@@ -8,12 +8,12 @@ import { AnimationProps } from '../../components/Animation/Animation';
 import { AttachmentProps } from '../../components/Attachment/Attachment';
 import { AvatarStylesProps } from '../../components/Avatar/Avatar';
 import { ButtonGroupProps } from '../../components/Button/ButtonGroup';
-import { ButtonStylesProps } from './components/Button/buttonStyles';
+import { ButtonStylesProps } from '../../components/Button/Button';
 import { ButtonContentStylesProps } from '../../components/Button/ButtonContent';
 import { ChatItemStylesProps } from '../../components/Chat/ChatItem';
 import { ChatMessageStylesProps } from '../../components/Chat/ChatMessage';
 import { ChatStylesProps } from '../../components/Chat/Chat';
-import { CheckboxStylesProps } from './components/Checkbox/checkboxStyles';
+import { CheckboxStylesProps } from '../../components/Checkbox/Checkbox';
 import { DividerProps } from '../../components/Divider/Divider';
 import { DropdownProps } from '../../components/Dropdown/Dropdown';
 import { EmbedProps } from '../../components/Embed/Embed';
@@ -24,14 +24,14 @@ import { FormProps } from '../../components/Form/Form';
 import { GridProps } from '../../components/Grid/Grid';
 import { HeaderDescriptionProps } from '../../components/Header/HeaderDescription';
 import { HeaderProps } from '../../components/Header/Header';
-import { IconStylesProps } from './components/Icon/iconStyles';
-import { ImageStylesProps } from './components/Image/imageStyles';
+import { IconStylesProps } from '../../components/Icon/Icon';
+import { ImageStylesProps } from '../../components/Image/Image';
 import { InputProps } from '../../components/Input/Input';
 import { ItemLayoutProps } from '../../components/ItemLayout/ItemLayout';
-import { LabelStylesProps } from './components/Label/labelStyles';
+import { LabelStylesProps } from '../../components/Label/Label';
 import { LayoutProps } from '../../components/Layout/Layout';
-import { ListStylesProps } from './components/List/listStyles';
-import { ListItemStylesProps } from './components/List/listItemStyles';
+import { ListStylesProps } from '../../components/List/List';
+import { ListItemStylesProps } from '../../components/List/ListItem';
 import { LoaderProps } from '../../components/Loader/Loader';
 import { MenuItemProps } from '../../components/Menu/MenuItem';
 import { MenuProps } from '../../components/Menu/Menu';
@@ -42,9 +42,9 @@ import { RadioGroupProps } from '../../components/RadioGroup/RadioGroup';
 import { ReactionGroupProps } from '../../components/Reaction/ReactionGroup';
 import { ReactionProps } from '../../components/Reaction/Reaction';
 import { SegmentProps } from '../../components/Segment/Segment';
-import { SliderStylesProps } from './components/Slider/sliderStyles';
-import { StatusStylesProps } from './components/Status/statusStyles';
-import { TextStylesProps } from './components/Text/textStyles';
+import { SliderStylesProps } from '../../components/Slider/Slider';
+import { StatusStylesProps } from '../../components/Status/Status';
+import { TextStylesProps } from '../../components/Text/Text';
 import { ToolbarDividerStylesProps } from '../../components/Toolbar/ToolbarDivider';
 import { ToolbarItemStylesProps } from '../../components/Toolbar/ToolbarItem';
 import { ToolbarCustomItemStylesProps } from '../../components/Toolbar/ToolbarCustomItem';
